Move linked aggregation definitions into their own file
